backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
redis==5.0.1
prometheus-client==0.21.0
pydantic-settings==2.6.1
python-multipart==0.0.9
